Fix size calculation error for PanelTray objects that are not the primary tray
Fixes LP: #750481

(bzr r1233.1.2)

+++ src/PanelTray.h
@@ -52,7 +52,7 @@ public:
   virtual void OnEntryAdded(unity::indicator::Entry::Ptr const& proxy);
 
 public:
-  int        _n_children;
+  guint8     _n_children;
   char     **_whitelist;
 protected:
   const gchar * GetName ();
